Split scan_controller
